@@ -1,11 +1,11 @@
 name: loading_overlay
 description: A modal progress indicator widget that fades in and out. Wrap around another widget to block access to widget during an async call.
-version: 0.2.1
+version: 0.3.0
 author: James Collins <james-collins@hotmail.co.nz>, Maurice McCabe <mmcc007@gmail.com>
 homepage: https://github.com/java-james/loading_overlay
 
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
